V2

+++ frontend/src/app/login/login.component.ts
@@ -1,8 +1,8 @@
-import {Component} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {FormsModule} from "@angular/forms";
-import {ApiService} from '../api.service';
-import {StorageService} from '../storage.service';
+import {StorageService} from "../storage.service";
+import {ApiService} from "../api.service";
 
 @Component({
   selector: 'app-login',
@@ -11,45 +11,66 @@ import {StorageService} from '../storage.service';
   templateUrl: './login.component.html',
   styleUrl: './login.component.scss'
 })
-export class LoginComponent {
+export class LoginComponent implements OnInit {
+  username = '';
+  password = '';
+  passwordConfirm = '';
+  isRegister = false;
+
   constructor(private _apiService: ApiService, private _storageService: StorageService) {
   }
 
+  ngOnInit(): void {
+  }
 
-  username: string = "";
-  password: string = "";
-
-  onLogin() {
+  onLogin(): void {
     if (this.username.trim() == "" || this.password.trim() == "") return;
 
     this._apiService.loginPost(this.username, this.password).pipe().subscribe(
-        data => {
-          sessionStorage.setItem('token', data);
-          console.log("Login successful");
+      data => {
+        sessionStorage.setItem('token', data);
+        this._storageService.masterPassword = this.password;
 
-          window.location.href = "/vault";
-          this._storageService.masterPassword = this.password;
-        },
-        error => {
-          console.warn("Login failed");
+        this._storageService.setLoggedIn(true);
+
+        console.log("Login: Success");
+      },
+      error => {
+        switch (error.error.error.code) {
+          case 401:
+          case 404:
+            console.warn("Login: Invalid username or password");
+            break;
+          default:
+            console.warn("Login: Unknown error");
+            break;
         }
+      }
     );
   }
 
-  onRegister() {
-    if (this.username.trim() == "" || this.password.trim() == "") return;
+  onRegister(): void {
+    if (this.username.trim() == "" || this.password.trim() == "" || this.passwordConfirm.trim() == "") return;
+    if (this.password != this.passwordConfirm) return;
 
     this._apiService.registerPost(this.username, this.password).pipe().subscribe(
-        data => {
-          sessionStorage.setItem('token', data);
-          console.log("Register successful");
+      data => {
+        sessionStorage.setItem('token', data);
+        this._storageService.masterPassword = this.password;
+        this._storageService.setLoggedIn(true);
 
-          window.location.href = "/vault";
-          this._storageService.masterPassword = this.password;
-        },
-        error => {
-          console.warn("Register failed");
+        console.log("Register: Success");
+      },
+      error => {
+        switch (error.status) {
+          case 409:
+            console.warn("Register: Username already exists");
+            break;
+          default:
+            console.warn("Register: Unknown error");
+            break;
         }
+      }
     );
   }
 }
